simplify phase3 validation setup (#1487)
